% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{element_grob.theme_border}
\alias{element_grob.theme_border}
\title{Title}
\usage{
element_grob.theme_border(
  element,
  x = 0,
  y = 0,
  width = 1,
  height = 1,
  type = NULL,
  colour = NULL,
  size = NULL,
  linetype = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
