% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{len0_null}
\alias{len0_null}
\title{Code duplicated from ggplot2 source (not exposed to wider namespace) for convenience}
\usage{
len0_null(x)
}
\arguments{
\item{x}{}
}
\description{
Code duplicated from ggplot2 source (not exposed to wider namespace) for convenience
}
