% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{make_plot_data}
\alias{make_plot_data}
\title{Prepare data for plots}
\usage{
make_plot_data(depth, position, length)
}
\arguments{
\item{depth}{depth by skate}

\item{position}{position by station}

\item{length}{lengths}
}
\value{
list of data used for making plots
}
\description{
Prepare data for plots
}
