% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_plots}
\alias{make_plots}
\title{Make QAQC plots}
\usage{
make_plots(depth, chx, len3, testE, testO, StnLoc)
}
\arguments{
\item{depth}{depths of each skate of gear}

\item{chx}{check that start and ends are near skate depths}

\item{len3}{sablefish length data}

\item{testE}{even lengths}

\item{testO}{odd lengths}

\item{StnLoc}{station locations to compare with standard setting positions}
}
\value{
saved plots
}
\description{
Make QAQC plots
}
