% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{qaqc}
\alias{qaqc}
\title{Run QAQC of AFSC longline survey station}
\usage{
qaqc(channel, station)
}
\arguments{
\item{channel}{channel to access database}

\item{station}{station number}
}
\value{
makes plots for qaqc
}
\description{
Run QAQC of AFSC longline survey station
}
