% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_border}
\alias{theme_border}
\title{Title}
\usage{
theme_border(
  type = c("left", "right", "bottom", "top", "none"),
  colour = "black",
  size = 1,
  linetype = 1
)
}
\arguments{
\item{linetype}{}
}
\description{
Title
}
