% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_L_border}
\alias{theme_L_border}
\title{For convenience: "L" (left + bottom) border}
\usage{
theme_L_border(colour = "black", size = 1, linetype = 1)
}
\arguments{
\item{linetype}{}
}
\description{
For convenience: "L" (left + bottom) border
}
