% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_plots}
\alias{make_plots}
\title{Make QAQC plots}
\usage{
make_plots(sta, depth, chx, lengths, lenE, lenO, StnLoc)
}
\arguments{
\item{sta}{station number}

\item{depth}{depths of each skate of gear}

\item{chx}{check that start and ends are near skate depths}

\item{lengths}{sablefish length data}

\item{lenE}{even lengths}

\item{lenO}{odd lengths}

\item{StnLoc}{station locations to compare with standard setting positions}
}
\value{
saved plots
}
\description{
Make QAQC plots
}
